ingest_function_source/requirements.txt
functions-framework==3.*
google-cloud-discoveryengine